+++ libs/components/core/src/lib/modules/screen-reader-label/fixtures/screen-reader-label.directive.fixture.ts
@@ -1,0 +1,18 @@
+import { CommonModule } from '@angular/common';
+import { Component, Input } from '@angular/core';
+
+import { SkyScreenReaderLabelDirective } from '../screen-reader-label.directive';
+
+@Component({
+  selector: 'sky-screen-reader-label-directive-test',
+  templateUrl: './screen-reader-label.directive.fixture.html',
+  standalone: true,
+  imports: [CommonModule, SkyScreenReaderLabelDirective],
+})
+export class ScreenReaderLabelFixtureComponent {
+  @Input()
+  public createLabel1 = false;
+
+  @Input()
+  public createLabel2?: boolean;
+}