+++ libs/components/lists/src/lib/modules/sort/sort.component.spec.ts
@@ -5,6 +5,7 @@ import {
   tick,
 } from '@angular/core/testing';
 import { expect, expectAsync } from '@skyux-sdk/testing';
+import { SkyContentInfoProvider } from '@skyux/core';
 import {
   SkyTheme,
   SkyThemeMode,
@@ -22,6 +23,7 @@ import { SkySortModule } from './sort.module';
 describe('Sort component', () => {
   let fixture: ComponentFixture<SortTestComponent>;
   let component: SortTestComponent;
+  let contentInfo: SkyContentInfoProvider;
   let mockThemeSvc: {
     settingsChange: BehaviorSubject<SkyThemeSettingsChange>;
   };
@@ -45,11 +47,14 @@ describe('Sort component', () => {
           provide: SkyThemeService,
           useValue: mockThemeSvc,
         },
+        SkyContentInfoProvider,
       ],
     });
 
     fixture = TestBed.createComponent(SortTestComponent);
     component = fixture.componentInstance;
+
+    contentInfo = TestBed.inject(SkyContentInfoProvider);
   });
 
   function getDropdownButtonEl(): HTMLElement | null {
@@ -117,6 +122,31 @@ describe('Sort component', () => {
     const dropdownButtonEl = getDropdownButtonEl();
     expect(dropdownButtonEl?.getAttribute('aria-label')).toBe('Test label');
   }));
+
+  it('should use the content info provider for aria label when applicable', () => {
+    contentInfo.patchInfo({
+      descriptor: { value: 'constituents', type: 'text' },
+    });
+    fixture.detectChanges();
+
+    const dropdownButtonEl = getDropdownButtonEl();
+    expect(dropdownButtonEl?.getAttribute('aria-label')).toBe(
+      'Sort constituents'
+    );
+  });
+
+  it('should not use the content info provider for aria label when overwritten', () => {
+    contentInfo.patchInfo({
+      descriptor: { value: 'constituents', type: 'text' },
+    });
+    component.ariaLabel = 'Overwritten label';
+    fixture.detectChanges();
+
+    const dropdownButtonEl = getDropdownButtonEl();
+    expect(dropdownButtonEl?.getAttribute('aria-label')).toBe(
+      'Overwritten label'
+    );
+  });
 
   it('changes active item on click and emits proper event', fakeAsync(() => {
     fixture.detectChanges();