+++ libs/components/lookup/src/lib/modules/search/search.component.spec.ts
@@ -8,7 +8,11 @@ import {
 import { By } from '@angular/platform-browser';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { expect, expectAsync } from '@skyux-sdk/testing';
-import { SkyMediaBreakpoints, SkyMediaQueryService } from '@skyux/core';
+import {
+  SkyContentInfoProvider,
+  SkyMediaBreakpoints,
+  SkyMediaQueryService,
+} from '@skyux/core';
 import { MockSkyMediaQueryService } from '@skyux/core/testing';
 import {
   SkyTheme,
@@ -28,6 +32,7 @@ describe('Search component', () => {
   let fixture: ComponentFixture<SearchTestComponent>;
   let component: SearchTestComponent;
   let element: DebugElement;
+  let contentInfoProvider: SkyContentInfoProvider;
   let mockMediaQueryService: MockSkyMediaQueryService;
   let mockThemeSvc: { settingsChange: BehaviorSubject<SkyThemeSettingsChange> };
 
@@ -56,6 +61,7 @@ describe('Search component', () => {
           provide: SkyThemeService,
           useValue: mockThemeSvc,
         },
+        SkyContentInfoProvider,
       ],
     });
 
@@ -71,6 +77,8 @@ describe('Search component', () => {
     }).createComponent(SearchTestComponent);
     component = fixture.componentInstance;
     element = fixture.debugElement as DebugElement;
+
+    contentInfoProvider = TestBed.inject(SkyContentInfoProvider);
   });
 
   afterEach(() => {
@@ -533,6 +541,35 @@ describe('Search component', () => {
       await fixture.whenStable();
       await expectAsync(fixture.nativeElement).toBeAccessible();
       expect(getInput().attributes['aria-label']).toBe('Test label');
+      expect(getInput().attributes['aria-labelledby']).toBeUndefined();
+    });
+
+    it('should be accessible using default theme at wide and small breakpoints (ariaLabel: "Search constituents" - via content info, ariaLabelledBy: undefined)', async () => {
+      contentInfoProvider.patchInfo({
+        descriptor: { value: 'constituents', type: 'text' },
+      });
+      fixture.detectChanges();
+      await fixture.whenStable();
+      await expectAsync(fixture.nativeElement).toBeAccessible();
+      setInput('foo bar');
+      await fixture.whenStable();
+      await expectAsync(fixture.nativeElement).toBeAccessible();
+      expect(getInput().attributes['aria-label']).toBe('Search constituents');
+      expect(getInput().attributes['aria-labelledby']).toBeUndefined();
+    });
+
+    it('should be accessible using default theme at wide and small breakpoints (ariaLabel: "Overwritten label" - overwriting content info, ariaLabelledBy: undefined)', async () => {
+      contentInfoProvider.patchInfo({
+        descriptor: { value: 'constituents', type: 'text' },
+      });
+      component.ariaLabel = 'Overwritten label';
+      fixture.detectChanges();
+      await fixture.whenStable();
+      await expectAsync(fixture.nativeElement).toBeAccessible();
+      setInput('foo bar');
+      await fixture.whenStable();
+      await expectAsync(fixture.nativeElement).toBeAccessible();
+      expect(getInput().attributes['aria-label']).toBe('Overwritten label');
       expect(getInput().attributes['aria-labelledby']).toBeUndefined();
     });