+++ libs/components/popovers/src/lib/modules/dropdown/dropdown.component.ts
@@ -2,6 +2,7 @@ import {
   ChangeDetectionStrategy,
   ChangeDetectorRef,
   Component,
+  DestroyRef,
   ElementRef,
   EnvironmentInjector,
   Input,
@@ -179,10 +180,6 @@ export class SkyDropdownComponent implements OnInit, OnDestroy {
     }
   }
 
-  #contentInfoProvider = inject(SkyContentInfoProvider, {
-    optional: true,
-  });
-
   public isMouseEnter = false;
 
   public isVisible = false;
@@ -210,6 +207,8 @@ export class SkyDropdownComponent implements OnInit, OnDestroy {
     return this.#_triggerButton;
   }
 
+  protected destroyRef = inject(DestroyRef);
+
   #affixer: SkyAffixer | undefined;
   #overlay: SkyOverlayInstance | undefined;
   #ngUnsubscribe = new Subject<void>();
@@ -221,6 +220,9 @@ export class SkyDropdownComponent implements OnInit, OnDestroy {
   readonly #overlayService = inject(SkyOverlayService);
   readonly #themeSvc = inject(SkyThemeService, { optional: true });
   readonly #zIndex = inject(SKY_STACKING_CONTEXT, { optional: true })?.zIndex;
+  readonly #contentInfoProvider = inject(SkyContentInfoProvider, {
+    optional: true,
+  });
 
   #_buttonStyle = DEFAULT_BUTTON_STYLE;
   #_buttonType = DEFAULT_BUTTON_TYPE;